Move mongo store to a separate package
