+++ mongo/mongo_test.go
@@ -1,39 +1,41 @@
-package sessions
+package mongo
 
 import (
 	"testing"
 
+	"github.com/gin-contrib/sessions"
+	"github.com/gin-contrib/sessions/tester"
 	mgo "gopkg.in/mgo.v2"
 )
 
 const mongoTestServer = "localhost:27017"
 
-var newMongoStore = func(_ *testing.T) Store {
+var newStore = func(_ *testing.T) sessions.Store {
 	session, err := mgo.Dial(mongoTestServer)
 	if err != nil {
 		panic(err)
 	}
 
 	c := session.DB("test").C("sessions")
-	return NewMongoStore(c, 3600, true, []byte("secret"))
+	return NewStore(c, 3600, true, []byte("secret"))
 }
 
 func TestMongo_SessionGetSet(t *testing.T) {
-	sessionGetSet(t, newMongoStore)
+	tester.GetSet(t, newStore)
 }
 
 func TestMongo_SessionDeleteKey(t *testing.T) {
-	sessionDeleteKey(t, newMongoStore)
+	tester.DeleteKey(t, newStore)
 }
 
 func TestMongo_SessionFlashes(t *testing.T) {
-	sessionFlashes(t, newMongoStore)
+	tester.Flashes(t, newStore)
 }
 
 func TestMongo_SessionClear(t *testing.T) {
-	sessionClear(t, newMongoStore)
+	tester.Clear(t, newStore)
 }
 
 func TestMongo_SessionOptions(t *testing.T) {
-	sessionOptions(t, newMongoStore)
+	tester.Options(t, newStore)
 }
